+++ packages/data/schema/system/src/model/SSS/file_and_storage/index.ts
@@ -1,0 +1,3 @@
+export * from './file_and_storage.interfaces';
+export { default as FileConfigModel, FileConfigModel as FileConfig } from './file_config.schema';
+export { default as StorageConfigModel, StorageConfigModel as StorageConfig } from './storage_config.schema';
